pyproject.toml
[project]
name = "langchain-promt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "langchain>=1.2.4",
    "langchain-core>=1.2.7",
    "langchain-openai>=1.1.7",
    "numpy>=2.4.1",
    "openai>=2.15.0",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "streamlit>=1.53.0",
    "uvicorn[standard]>=0.40.0",
]
